backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pytest==7.4.3
gunicorn==21.2.0
eventlet==0.33.3
apispec==6.3.0
apispec-webframeworks==0.5.2
marshmallow==3.20.1
Flask-Limiter==3.5.0
flask-apispec==0.11.1









